move batch analysis logic to service to use in graphs
