+++ src/app/motor-inventory/batch-analysis/batch-analysis-banner/batch-analysis-banner.component.ts
@@ -41,5 +41,6 @@ export class BatchAnalysisBannerComponent implements OnInit {
 
   saveSettings(){
     this.batchAnalysisService.batchAnalysisSettings.next(this.batchAnalysisSettings);
+    this.batchAnalysisService.updateReplaceMotorDecision();
   }
 }
