+++ src/app/motor-inventory/batch-analysis/batch-analysis-graphs/batch-analysis-graphs.component.html
@@ -1,0 +1,1 @@
+<p>batch-analysis-graphs works!</p>
