+++ src/app/motor-inventory/batch-analysis/batch-analysis-graphs/batch-analysis-graphs.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-batch-analysis-graphs',
+  templateUrl: './batch-analysis-graphs.component.html',
+  styleUrls: ['./batch-analysis-graphs.component.css']
+})
+export class BatchAnalysisGraphsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
