+++ src/app/motor-inventory/batch-analysis/batch-analysis.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { BatchAnalysisService } from './batch-analysis.service';
 import { Subscription } from 'rxjs';
+import { MotorInventoryService } from '../motor-inventory.service';
 
 @Component({
   selector: 'app-batch-analysis',
@@ -11,16 +12,24 @@ export class BatchAnalysisComponent implements OnInit {
 
   selectedTab: string;
   selectedTabSub: Subscription;
-  constructor(private batchAnalysisService: BatchAnalysisService) { }
+  filterInventorySummarySub: Subscription;
+  batchAnalysisSettingsSub: Subscription;
+  constructor(private batchAnalysisService: BatchAnalysisService, private motorInventoryService: MotorInventoryService) { }
 
   ngOnInit(): void {
     this.selectedTabSub = this.batchAnalysisService.selectedTab.subscribe(val => {
       this.selectedTab = val;
-    })
+    });
+
+    this.filterInventorySummarySub = this.motorInventoryService.filterInventorySummary.subscribe(() => {
+      this.batchAnalysisService.setBatchAnalysisDataItems();
+    });
   }
 
-  ngOnDestroy(){
+  ngOnDestroy() {
     this.selectedTabSub.unsubscribe();
+    this.batchAnalysisSettingsSub.unsubscribe();
+    this.filterInventorySummarySub.unsubscribe();
   }
 }
